refactor: use setuptools_scm to set version
refactor: split widgets and leaflet tools into separate modules
