+++ IS2view/version.py
@@ -8,6 +8,6 @@ from pkg_resources import get_distribution
 # get semantic version from setuptools-scm
 version = get_distribution("IS2view").version
 # append "v" before the version
-full_version = "v{0}".format(version)
+full_version = f"v{version}"
 # get project name
 project_name = get_distribution("IS2view").project_name
