--- doc/source/api_reference/IS2view.rst
@@ -1,29 +1,0 @@
-=======
-IS2view
-=======
-
-`Jupyter notebook <https://jupyter.org/>`_, `user interface <https://ipywidgets.readthedocs.io/en/latest/>`_ and plotting tools for visualizing `rioxarray <https://corteva.github.io/rioxarray/stable/>`_ variables on `ipyleaflet <https://ipyleaflet.readthedocs.io/en/latest/>`_ maps
-
-`Source code`__
-
-.. __: https://github.com/tsutterley/IS2view/blob/main/IS2view/IS2view.py
-
-General Attributes and Methods
-==============================
-
-.. autoclass:: IS2view.IS2view.widgets
-   :members:
-
-.. autoclass:: IS2view.IS2view.leaflet
-   :members:
-
-.. autofunction:: IS2view.IS2view.image_service_layer
-
-.. autoclass:: IS2view.IS2view.LeafletMap
-   :members:
-
-.. autoclass:: IS2view.IS2view.TimeSeries
-   :members:
-
-.. autoclass:: IS2view.IS2view.Transect
-   :members:
