+++ doc/source/api_reference/tools.rst
@@ -1,0 +1,15 @@
+=====
+tools
+=====
+
+`User interface <https://ipywidgets.readthedocs.io/en/latest/>`_ tools for `Jupyter notebook <https://jupyter.org/>`_
+
+`Source code`__
+
+.. __: https://github.com/tsutterley/IS2view/blob/main/IS2view/tools.py
+
+General Attributes and Methods
+==============================
+
+.. autoclass:: IS2view.tools.widgets
+   :members:
