+++ doc/source/api_reference/visualization.rst
@@ -1,0 +1,26 @@
+=============
+visualization
+=============
+
+Plotting tools for visualizing `rioxarray <https://corteva.github.io/rioxarray/stable/>`_ variables on `ipyleaflet <https://ipyleaflet.readthedocs.io/en/latest/>`_ maps
+
+`Source code`__
+
+.. __: https://github.com/tsutterley/IS2view/blob/main/IS2view/visualization.py
+
+General Attributes and Methods
+==============================
+
+.. autoclass:: IS2view.visualization.leaflet
+   :members:
+
+.. autofunction:: IS2view.visualization.image_service_layer
+
+.. autoclass:: IS2view.visualization.LeafletMap
+   :members:
+
+.. autoclass:: IS2view.visualization.TimeSeries
+   :members:
+
+.. autoclass:: IS2view.visualization.Transect
+   :members:
